Moderately large refactor to support different MatrixClient implementations. This will make integration testing a little easier to do by allowing for a mocked implementation to be injected.
The refactor includes:

 * `MessageData` structs no longer need to be cloned when being used as an argument to `MatrixClient.send_message(...)`.
 * Things related to the final `HttpResponse` (e.g. which response code to use) are mostly centralized in the `index(...)` function of the webhook_api crate.
 * Reduced duplication of code due to Sonarr/Radarr webhook differences.
 * `webhook_config` is now generic, expecting some `MatrixClient` implementation.
 * The facades no longer send Matrix messages themselves; they just generate the `MessageData` and the `index(...)` function sends the final message(s).

Reorganize some of the members of the matrix client crate.

Break the base functionality desired from the matrix client out to a separate trait.

Move some of the Matrix initialization stuff out to the main binary crate.

Move the default client logic (e.g. connecting and setting up the event loop) to a module within the matrix_client crate.

Rename initialization.rs to first_time_initialization.rs.
