+++ Cargo.toml
@@ -21,6 +21,7 @@ anyhow = "1.0.42"
 actix-web = {version = "4.0.0-beta.8"}
 log = "0.4"
 env_logger = "0.8"
+url = "2.2.2"
 tokio = { version = "1", features = ["rt", "sync"] }
 yarrbot_common = { path = "crates/common" }
 yarrbot_db = { path = "crates/db" }
