+++ crates/matrix_client/Cargo.toml
@@ -15,5 +15,6 @@ tokio = { version = "1.0", features = ["rt"] }
 futures = "0.3.16"
 uuid = { version = "= 0.8.2", features = ["v4"] }
 rand = { version = "0.8.4", features = ["small_rng"] }
+async-trait = "0.1.51"
 yarrbot_db = { path = "../db" }
 yarrbot_common = { path = "../common" }
